fix missing includes and incorrect signatures

+++ drm/sections/MaterialSection.h
@@ -21,7 +21,7 @@ class MaterialSection : public ResolveSection {
 public:
 	uint32_t realize(uint32_t sectionId, uint32_t unknown, uint32_t size, bool& alreadyLoaded) override;
 	// inherit allocate
-	void fill(uint32_t id, void* src, size_t size, size_t offset) override;
+	void fill(uint32_t id, void* src, uint32_t size, uint32_t offset) override;
 	void construct(uint32_t, void *) override;
 	void* getWrapped(uint32_t) override;
 	void* getBlob(uint32_t) override;
